requirements.txt
asgiref==3.8.1
Django==4.2.13
django-filter==24.2
djangorestframework==3.15.1
drf-yasg==1.21.7
psycopg2-binary==2.9.9
sqlparse==0.5.0
